shared/pyproject.toml
[project]
name = "loyalty-shared-library"
version = "0.1.0"
description = "Shared schemas, utilities, and data contracts for the Loyalty Program project."
dependencies = [
    "pydantic>=2.0",
    "python-jose[cryptography]>=3.3.0"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"